requirements.txt
###rasa_nlu
gevent==1.2.2

klein==17.10.0

hyperlink==17.3.1

boto3==1.5.20

typing==3.6.2

future==0.16.0

six==1.11.0

jsonschema==2.6.0

matplotlib==2.1.0

requests==2.18.4

tqdm==4.19.5

numpy==1.14.4

simplejson==3.13.2


###sklearn


spacy==2.0.11
scikit-learn==0.19.1

scipy==1.1.0

sklearn-crfsuite==0.3.6

###rasa_core
jsonpickle==0.9.6

six==1.11.0

redis==2.10.6

fakeredis==0.10.3

ruamel.yaml==0.15.37

graphviz==0.8.3

keras==2.1.6

tensorflow==1.8.0

h5py==2.7.1

apscheduler==3.5.1

ConfigArgParse==0.13.0

networkx==2.1

fbmessenger==5.2.0

pykwalify==1.6.0

coloredlogs==10.0

flask==1.0.2

flask_cors==3.0.4

scikit-learn==0.19.1

rasa_nlu~=0.12.3

slackclient==1.2.1

python-telegram-bot==10.1.0

twilio==6.14.0

mattermostwrapper==2.1

colorhash==1.0.2



## additional

pypandoc==1.4

rasa_core==0.10.1

zomatopy==1.0.10


